+++ src/client/app/widgets/index.js
@@ -22,10 +22,11 @@ module.exports.widgets = {
 
     // matrices
     matrix: require('./matrices/matrix'),
+    keyboard: require('./matrices/keyboard'),
+    // deprecated / hidden
     multitoggle: require('./matrices/multitoggle'),
     multipush: require('./matrices/multipush'),
     multifader: require('./matrices/multifader'),
-    keyboard: require('./matrices/keyboard'),
 
     // plots
     led: require('./plots/led'),
@@ -65,7 +66,7 @@ module.exports.categories = {
     'Sliders':['fader','knob', 'encoder', 'range'],
     'Buttons':['toggle','push','switch', 'dropdown'],
     'Pads':['xy','rgb','multixy'],
-    'Matrices':['matrix', 'multifader','multitoggle','multipush', 'keyboard'],
+    'Matrices':['matrix', 'keyboard'],
     'Plots':['plot','eq','visualizer','led', 'rgbled','meter','text', 'image', 'svg', 'frame'],
     'Containers':['panel','strip','modal', 'clone'],
     'Switchers':['switcher','crossfader'],